wait before moving
The converter checks for extern to be written asuming that is the last to happen. But "Waters ACQUITY FTN Postrun Report" is written to _INLET.INF afterwards.
If files are moved before that happens CHROM stays locked and cannot be moved and also the machine blocks and won't continue to next file. So I will just wait 30 sec for the file to be release (I hope)